Refactor CreatePackageDocs to accept a template path
The CreatePackageDocs function in docGeneration.go has been refactored to accept a template path as a parameter. This allows for flexibility in choosing the template file for generating the README files. The template path is now passed through the function call chain to generateReadmeForPackage and generateReadmeFromTemplate functions.

Changes made:

- Added templatePath parameter to the CreatePackageDocs function signature.
- Updated the handleDirectory function signature to include the templatePath parameter.
- Modified the processGoFiles function signature to include the templatePath parameter.
- Updated the generateReadmeForPackage function signature to include the templatePath parameter.
- Modified the generateReadmeFromTemplate function signature to include the templatePath parameter.
- Updated function calls in ExampleCreatePackageDocs and TestCreatePackageDocs to pass the templatePath parameter.
- Updated the UpdateDocs function in magefile.go to pass the templatePath parameter when calling CreatePackageDocs.

Note: The specific value for the templatePath parameter should be
provided based on the desired template file location.